refactor: ScreeningListActivity를 위한 UI 상태 클래스들을 한 패키지에 넣음
