+++ app/src/main/java/woowacourse/movie/view/activities/home/fragments/screeninglist/ScreeningListPresenter.kt
@@ -1,6 +1,7 @@
 package woowacourse.movie.view.activities.home.fragments.screeninglist
 
 import woowacourse.movie.repository.ScreeningRepository
+import woowacourse.movie.view.activities.home.fragments.screeninglist.uistates.ScreeningListUIState
 import woowacourse.movie.view.activities.home.fragments.screeninglist.uistates.TheatersUIState
 
 class ScreeningListPresenter(
