--- app/src/main/java/woowacourse/movie/view/activities/home/fragments/screeninglist/ScreeningListViewItemUIState.kt
@@ -1,3 +1,0 @@
-package woowacourse.movie.view.activities.home.fragments.screeninglist
-
-sealed class ScreeningListViewItemUIState
