+++ app/src/main/java/woowacourse/movie/view/activities/home/fragments/screeninglist/uistates/ScreeningListViewItemUIState.kt
@@ -1,0 +1,3 @@
+package woowacourse.movie.view.activities.home.fragments.screeninglist.uistates
+
+sealed class ScreeningListViewItemUIState
